scraper/requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
cloudscraper==1.2.71
dnspython==2.7.0
geographiclib==2.0
geopy==2.4.1
idna==3.10
lxml==5.3.0
numpy==2.1.3
pymongo==4.10.1
pyparsing==3.2.0
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
soupsieve==2.6
urllib3==2.2.3
wheel==0.44.0
